+++ CHANGELOG.md
@@ -19,6 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **BREAKING**: `OUDSBulletList.UnorderedAsset.free` case
 - **BREAKING**: `OUDSCheckboxItem(isOn:)`, `OUDSRadioItem(isOn:)`, `OUDSCheckboxItemIndeterminate(selection:)` inits
 - **BREAKING**: `OUDSTabBar(selected:count)` init
+### Added
+
+- `badge` on `toolbar` item (Orange-OpenSource/ouds-ios#1398)
 
 ### Fixed